+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to the GLB Viewer extension will be documented in this file.
 ### Changed
  - All core files are now in a separate repository: https://github.com/ohzinteractive/glb-viewer-core. This repository is now only a wrapper for the core files with some VS Code specific code.
 
+### Added
+ - Info panel now displays file size
+
 ## [2.13.0] - 2025-09-29
 
 ### Fixed